backend/land_registry/requirements.txt
django
djangorestframework
python-dotenv
mysqlclient